+++ src/lib.rs
@@ -1,6 +1,6 @@
 // TODO: Everything here and in the pulseq-rs crate should use f64!
 // Individual samples can go down to sub-microsecond scale while sequence durations
-// often are in seconds or minutes. Single-precision f32 is not sufficient in the
+// often are in seconds or minutes. Single-precision f64 is not sufficient in the
 // general case. It probably is enough for simulation with absolute times, but
 // errors accumulate too quickly when generating sequences -> use double precision!
 
@@ -23,26 +23,26 @@ pub fn load_pulseq<P: AsRef<Path>>(path: P) -> Result<Sequence, pulseq_rs::Error
 /// on which the public disseqt API is built upon
 trait Backend: Send {
     /// Return the FOV of the Sequence, if it is available
-    fn fov(&self) -> Option<(f32, f32, f32)>;
+    fn fov(&self) -> Option<(f64, f64, f64)>;
 
     /// Duration of the MRI sequence: no samples, blocks, etc. exist outside
     /// of the time range [0, duration()]
-    fn duration(&self) -> f32;
+    fn duration(&self) -> f64;
 
     /// Returns all events of the given type in the given duration.
     /// t_start is inclusive, t_end is exclusive. If a max_count is given and
     /// reached, there might be more events in the time span that are not returned.
-    fn events(&self, ty: EventType, t_start: f32, t_end: f32, max_count: usize) -> Vec<f32>;
+    fn events(&self, ty: EventType, t_start: f64, t_end: f64, max_count: usize) -> Vec<f64>;
 
     /// Returns the time range of the next encounter of the given type.
     /// If `t_start` is inside of a block, this block is not returned: only
     /// blocks **starting** after (or exactly on) `t_start` are considered.
     /// TODO: EventType should be the first parameter
-    fn encounter(&self, t_start: f32, ty: EventType) -> Option<(f32, f32)>;
+    fn encounter(&self, t_start: f64, ty: EventType) -> Option<(f64, f64)>;
 
     /// Samples the sequence at the given time points
-    fn sample(&self, time: &[f32]) -> Vec<Sample>;
+    fn sample(&self, time: &[f64]) -> Vec<Sample>;
 
     /// Integrates over the n-1 time intervalls given by the list of n time points.
-    fn integrate(&self, time: &[f32]) -> Vec<Moment>;
+    fn integrate(&self, time: &[f64]) -> Vec<Moment>;
 }
