+++ src/util.rs
@@ -1,7 +1,7 @@
 use std::ops::MulAssign;
 
 // Internally, a double precision is used, but interfaces are in single precision.
-// This is done to reduce accumulation of errors, but all other code still uses f32.
+// This is done to reduce accumulation of errors, but all other code still uses f64.
 
 pub struct Spin([f64; 3]);
 
@@ -10,17 +10,17 @@ impl Spin {
         Self([0.0, 0.0, 1.0])
     }
 
-    pub fn angle(&self) -> f32 {
+    pub fn angle(&self) -> f64 {
         // Normalize because error can build up during rotations
-        (self.0[2] / self.norm()).acos() as f32
+        (self.0[2] / self.norm()).acos() as f64
     }
 
-    pub fn phase(&self) -> f32 {
+    pub fn phase(&self) -> f64 {
         // We want the phase of the applied rotation, not of the spin itself
-        let tmp = f64::atan2(self.0[1], self.0[0]) as f32 + std::f32::consts::FRAC_PI_2;
+        let tmp = f64::atan2(self.0[1], self.0[0]) as f64 + std::f64::consts::FRAC_PI_2;
         // Map to the range [0, 2*pi]
         if tmp < 0.0 {
-            tmp + std::f32::consts::TAU
+            tmp + std::f64::consts::TAU
         } else {
             tmp
         }
@@ -34,7 +34,7 @@ impl Spin {
 pub struct Rotation([[f64; 3]; 3]);
 
 impl Rotation {
-    pub fn new(angle: f32, phase: f32) -> Self {
+    pub fn new(angle: f64, phase: f64) -> Self {
         let angle = angle as f64;
         let phase = phase as f64;
         Self([
@@ -74,8 +74,8 @@ mod tests {
     #[test]
     fn random_rot() {
         for _ in 0..1000 {
-            let angle = rand::random::<f32>() * std::f32::consts::PI;
-            let phase = rand::random::<f32>() * std::f32::consts::TAU;
+            let angle = rand::random::<f64>() * std::f64::consts::PI;
+            let phase = rand::random::<f64>() * std::f64::consts::TAU;
 
             let mut spin = Spin::relaxed();
             spin *= Rotation::new(angle, phase);
@@ -88,13 +88,13 @@ mod tests {
     #[test]
     fn random_multi_rot() {
         for _ in 0..1000 {
-            let angle = rand::random::<f32>() * std::f32::consts::PI;
-            let phase = rand::random::<f32>() * std::f32::consts::TAU;
+            let angle = rand::random::<f64>() * std::f64::consts::PI;
+            let phase = rand::random::<f64>() * std::f64::consts::TAU;
 
             let mut spin = Spin::relaxed();
             let subsamples = rand::random::<u32>() % 100 + 1;
             for _ in 0..subsamples {
-                spin *= Rotation::new(angle / subsamples as f32, phase);
+                spin *= Rotation::new(angle / subsamples as f64, phase);
             }
 
             check!((spin.angle() - angle).abs() < 1e-6);
